Remove WebSocket-projects. Move everything into the corresponding Core packages.
