+++ ExRam.Gremlinq.sln
@@ -9,8 +9,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExRam.Gremlinq.Core.Tests",
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExRam.Gremlinq.Providers.CosmosDb", "src\ExRam.Gremlinq.Providers.CosmosDb\ExRam.Gremlinq.Providers.CosmosDb.csproj", "{1CFD6C6F-137F-4402-ACF2-A7775824282D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExRam.Gremlinq.Providers.WebSocket", "src\ExRam.Gremlinq.Providers.WebSocket\ExRam.Gremlinq.Providers.WebSocket.csproj", "{8A621708-9F8A-4727-9D96-71B07CAF61A2}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{4E6D1C9D-C1F7-43CF-99F7-9D610B3D1989}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExRam.Gremlinq.Providers.CosmosDb.Tests", "test\ExRam.Gremlinq.Providers.CosmosDb.Tests\ExRam.Gremlinq.Providers.CosmosDb.Tests.csproj", "{E0F57FD2-0666-42B5-95ED-14D155A47F38}"
@@ -40,8 +38,6 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "AspNet", "AspNet", "{179A3F19-19B0-4DE2-B843-83AF4241244A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExRam.Gremlinq.Providers.GremlinServer.AspNet", "src\ExRam.Gremlinq.Providers.GremlinServer.AspNet\ExRam.Gremlinq.Providers.GremlinServer.AspNet.csproj", "{A3E92EDA-3FAC-4B65-BE3E-106A1BB1167C}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExRam.Gremlinq.Providers.WebSocket.AspNet", "src\ExRam.Gremlinq.Providers.WebSocket.AspNet\ExRam.Gremlinq.Providers.WebSocket.AspNet.csproj", "{B15DCF7F-D3C8-4ED2-ADED-0743D5CF2377}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExRam.Gremlinq.Providers.CosmosDb.AspNet", "src\ExRam.Gremlinq.Providers.CosmosDb.AspNet\ExRam.Gremlinq.Providers.CosmosDb.AspNet.csproj", "{B3DD7F6D-49DE-42CE-97BA-F90F9820D126}"
 EndProject
@@ -75,10 +71,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gremlinq.Providers.Neptune.
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gremlinq.Providers.Neptune", "meta\Gremlinq.Providers.Neptune\Gremlinq.Providers.Neptune.csproj", "{C22A5822-96FE-4A0F-A093-80DD605ED1E5}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gremlinq.Providers.WebSocket.AspNet", "meta\Gremlinq.Providers.WebSocket.AspNet\Gremlinq.Providers.WebSocket.AspNet.csproj", "{60D9CE5D-C808-4DE0-BC4E-164BF9B30310}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Gremlinq.Providers.WebSocket", "meta\Gremlinq.Providers.WebSocket\Gremlinq.Providers.WebSocket.csproj", "{2DEAF295-8965-4660-93FE-8AD3ED985CB3}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BenchmarkRunner", "src\Benchmarks\BenchmarkRunner\BenchmarkRunner.csproj", "{08B239B3-1A9A-4D29-AF18-50B8EEAFF97A}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExRam.Gremlinq.Providers.Core", "src\ExRam.Gremlinq.Providers.Core\ExRam.Gremlinq.Providers.Core.csproj", "{358A8F94-4AEC-4A56-8568-96814BB019A0}"
@@ -111,10 +103,6 @@ Global
 		{1CFD6C6F-137F-4402-ACF2-A7775824282D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1CFD6C6F-137F-4402-ACF2-A7775824282D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1CFD6C6F-137F-4402-ACF2-A7775824282D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8A621708-9F8A-4727-9D96-71B07CAF61A2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8A621708-9F8A-4727-9D96-71B07CAF61A2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8A621708-9F8A-4727-9D96-71B07CAF61A2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8A621708-9F8A-4727-9D96-71B07CAF61A2}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E0F57FD2-0666-42B5-95ED-14D155A47F38}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{E0F57FD2-0666-42B5-95ED-14D155A47F38}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E0F57FD2-0666-42B5-95ED-14D155A47F38}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -155,10 +143,6 @@ Global
 		{A3E92EDA-3FAC-4B65-BE3E-106A1BB1167C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A3E92EDA-3FAC-4B65-BE3E-106A1BB1167C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A3E92EDA-3FAC-4B65-BE3E-106A1BB1167C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B15DCF7F-D3C8-4ED2-ADED-0743D5CF2377}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B15DCF7F-D3C8-4ED2-ADED-0743D5CF2377}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B15DCF7F-D3C8-4ED2-ADED-0743D5CF2377}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B15DCF7F-D3C8-4ED2-ADED-0743D5CF2377}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B3DD7F6D-49DE-42CE-97BA-F90F9820D126}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B3DD7F6D-49DE-42CE-97BA-F90F9820D126}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B3DD7F6D-49DE-42CE-97BA-F90F9820D126}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -219,14 +203,6 @@ Global
 		{C22A5822-96FE-4A0F-A093-80DD605ED1E5}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C22A5822-96FE-4A0F-A093-80DD605ED1E5}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C22A5822-96FE-4A0F-A093-80DD605ED1E5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{60D9CE5D-C808-4DE0-BC4E-164BF9B30310}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{60D9CE5D-C808-4DE0-BC4E-164BF9B30310}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{60D9CE5D-C808-4DE0-BC4E-164BF9B30310}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{60D9CE5D-C808-4DE0-BC4E-164BF9B30310}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2DEAF295-8965-4660-93FE-8AD3ED985CB3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2DEAF295-8965-4660-93FE-8AD3ED985CB3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2DEAF295-8965-4660-93FE-8AD3ED985CB3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2DEAF295-8965-4660-93FE-8AD3ED985CB3}.Release|Any CPU.Build.0 = Release|Any CPU
 		{08B239B3-1A9A-4D29-AF18-50B8EEAFF97A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{08B239B3-1A9A-4D29-AF18-50B8EEAFF97A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{08B239B3-1A9A-4D29-AF18-50B8EEAFF97A}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -269,7 +245,6 @@ Global
 		{D4BB3324-E416-4C41-9FCB-1C1FDA5FE7E4} = {4E6D1C9D-C1F7-43CF-99F7-9D610B3D1989}
 		{6E351A4C-C1E6-4EEB-A956-3E241CB377A9} = {179A3F19-19B0-4DE2-B843-83AF4241244A}
 		{A3E92EDA-3FAC-4B65-BE3E-106A1BB1167C} = {179A3F19-19B0-4DE2-B843-83AF4241244A}
-		{B15DCF7F-D3C8-4ED2-ADED-0743D5CF2377} = {179A3F19-19B0-4DE2-B843-83AF4241244A}
 		{B3DD7F6D-49DE-42CE-97BA-F90F9820D126} = {179A3F19-19B0-4DE2-B843-83AF4241244A}
 		{7C211878-D55B-444E-AFD0-4095A5B3573B} = {179A3F19-19B0-4DE2-B843-83AF4241244A}
 		{69C738C8-37D2-41DB-A8DE-00E38B22BCA6} = {179A3F19-19B0-4DE2-B843-83AF4241244A}
@@ -284,8 +259,6 @@ Global
 		{7BAC7812-99B6-4B79-9480-E897C84E56C6} = {E944FDE9-B35D-4B16-B7A1-BF3644F5B7FA}
 		{ABF1F75B-57A8-46B9-B4EE-FE55B92608DF} = {E944FDE9-B35D-4B16-B7A1-BF3644F5B7FA}
 		{C22A5822-96FE-4A0F-A093-80DD605ED1E5} = {E944FDE9-B35D-4B16-B7A1-BF3644F5B7FA}
-		{60D9CE5D-C808-4DE0-BC4E-164BF9B30310} = {E944FDE9-B35D-4B16-B7A1-BF3644F5B7FA}
-		{2DEAF295-8965-4660-93FE-8AD3ED985CB3} = {E944FDE9-B35D-4B16-B7A1-BF3644F5B7FA}
 		{08B239B3-1A9A-4D29-AF18-50B8EEAFF97A} = {2CD9652B-1596-4A02-AF9B-6CE149AB3F65}
 		{0426F77B-0949-438C-864C-312DDA530CBC} = {E944FDE9-B35D-4B16-B7A1-BF3644F5B7FA}
 		{4B7289D1-5E6D-45BA-B811-8CA87E2DACE2} = {4E6D1C9D-C1F7-43CF-99F7-9D610B3D1989}
