--- meta/Gremlinq.Providers.WebSocket.AspNet/Gremlinq.Providers.WebSocket.AspNet.csproj
@@ -1,3 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-</Project>
