--- meta/Gremlinq.Providers.WebSocket/Gremlinq.Providers.WebSocket.csproj
@@ -1,3 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-</Project>
