+++ src/ExRam.Gremlinq.Providers.Core.AspNet/Extensions/ProviderSetupExtensions.cs
@@ -1,5 +1,9 @@
-﻿using ExRam.Gremlinq.Core;
+﻿using System.Net.WebSockets;
+using ExRam.Gremlinq.Core;
 using ExRam.Gremlinq.Core.AspNet;
+using ExRam.Gremlinq.Providers.WebSocket;
+using Gremlin.Net.Driver;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace ExRam.Gremlinq.Providers.Core.AspNet
@@ -21,6 +25,51 @@ namespace ExRam.Gremlinq.Providers.Core.AspNet
             }
 
             public TConfigurator Transform(TConfigurator configurator) => _extraConfiguration(configurator, _gremlinqSection, _providerSection);
+        }
+
+        private sealed class ConnectionPoolSettingsGremlinClientFactory : IGremlinClientFactory
+        {
+            private readonly IGremlinClientFactory _factory;
+            private readonly IConfigurationSection _section;
+
+            public ConnectionPoolSettingsGremlinClientFactory(IGremlinClientFactory factory, IConfigurationSection section)
+            {
+                _factory = factory;
+                _section = section;
+            }
+
+            public IGremlinClient Create(GremlinServer gremlinServer, IMessageSerializer messageSerializer, ConnectionPoolSettings connectionPoolSettings, Action<ClientWebSocketOptions> webSocketConfiguration, string? sessionId = null)
+            {
+                if (int.TryParse(_section[$"{nameof(ConnectionPoolSettings.MaxInProcessPerConnection)}"], out var maxInProcessPerConnection))
+                    connectionPoolSettings.MaxInProcessPerConnection = maxInProcessPerConnection;
+
+                if (int.TryParse(_section[$"{nameof(ConnectionPoolSettings.PoolSize)}"], out var poolSize))
+                    connectionPoolSettings.PoolSize = poolSize;
+
+                return _factory.Create(gremlinServer, messageSerializer, connectionPoolSettings, webSocketConfiguration, sessionId);
+            }
+        }
+
+        public static ProviderSetup<TConfigurator> ConfigureWebSocket<TConfigurator>(this ProviderSetup<TConfigurator> setup)
+           where TConfigurator : IWebSocketProviderConfigurator<TConfigurator>
+        {
+            return setup
+                .Configure((configurator, providerSection) =>
+                {
+                    var authenticationSection = providerSection.GetSection("Authentication");
+                    var connectionPoolSection = providerSection.GetSection("ConnectionPool");
+
+                    if (providerSection["Uri"] is { } uri)
+                        configurator = configurator.At(uri);
+
+                    configurator
+                        .ConfigureClientFactory(factory => new ConnectionPoolSettingsGremlinClientFactory(factory, connectionPoolSection));
+
+                    if (authenticationSection["Username"] is { } username && authenticationSection["Password"] is { } password)
+                        configurator = configurator.AuthenticateBy(username, password);
+
+                    return configurator;
+                });
         }
 
         public static ProviderSetup<TConfigurator> Configure<TConfigurator>(this ProviderSetup<TConfigurator> setup)
