+++ src/ExRam.Gremlinq.Providers.Core/WebSocket/DefaultMessageSerializer.cs
