+++ src/ExRam.Gremlinq.Providers.Core/WebSocket/IWebSocketProviderConfigurator.cs
