+++ src/ExRam.Gremlinq.Providers.Core/WebSocket/WebSocketProviderConfigurator.cs
