+++ src/ExRam.Gremlinq.Providers.CosmosDb.AspNet/ExRam.Gremlinq.Providers.CosmosDb.AspNet.csproj
@@ -6,7 +6,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ExRam.Gremlinq.Providers.CosmosDb\ExRam.Gremlinq.Providers.CosmosDb.csproj" />
-    <ProjectReference Include="..\ExRam.Gremlinq.Providers.WebSocket.AspNet\ExRam.Gremlinq.Providers.WebSocket.AspNet.csproj" />
+    <ProjectReference Include="..\ExRam.Gremlinq.Providers.Core.AspNet\ExRam.Gremlinq.Providers.Core.AspNet.csproj" />
   </ItemGroup>
 
 </Project>
