+++ src/ExRam.Gremlinq.Providers.Neptune.AspNet/ExRam.Gremlinq.Providers.Neptune.AspNet.csproj
@@ -6,7 +6,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\ExRam.Gremlinq.Providers.Neptune\ExRam.Gremlinq.Providers.Neptune.csproj" />
-    <ProjectReference Include="..\ExRam.Gremlinq.Providers.WebSocket.AspNet\ExRam.Gremlinq.Providers.WebSocket.AspNet.csproj" />
+    <ProjectReference Include="..\ExRam.Gremlinq.Providers.Core.AspNet\ExRam.Gremlinq.Providers.Core.AspNet.csproj" />
   </ItemGroup>
 
 </Project>
