+++ src/ExRam.Gremlinq.Providers.Neptune/ExRam.Gremlinq.Providers.Neptune.csproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ExRam.Gremlinq.Providers.WebSocket\ExRam.Gremlinq.Providers.WebSocket.csproj" />
+    <ProjectReference Include="..\ExRam.Gremlinq.Providers.Core\ExRam.Gremlinq.Providers.Core.csproj" />
   </ItemGroup>
 
 </Project>
