--- src/ExRam.Gremlinq.Providers.WebSocket.AspNet/ExRam.Gremlinq.Providers.WebSocket.AspNet.csproj
@@ -1,12 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <Description>ASP.NET Core support package for ExRam.Gremlinq.Providers.WebSocket.</Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ExRam.Gremlinq.Providers.Core.AspNet\ExRam.Gremlinq.Providers.Core.AspNet.csproj" />
-    <ProjectReference Include="..\ExRam.Gremlinq.Providers.WebSocket\ExRam.Gremlinq.Providers.WebSocket.csproj" />
-  </ItemGroup>
-
-</Project>
