--- src/ExRam.Gremlinq.Providers.WebSocket/ExRam.Gremlinq.Providers.WebSocket.csproj
@@ -1,11 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <Description>WebSocket provider binding for ExRam.Gremlinq.</Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ExRam.Gremlinq.Providers.Core\ExRam.Gremlinq.Providers.Core.csproj" />
-  </ItemGroup>
-
-</Project>
