+++ test/ExRam.Gremlinq.PublicApi.Tests/ExRam.Gremlinq.PublicApi.Tests.csproj
@@ -14,7 +14,6 @@
     <ProjectReference Include="..\..\src\ExRam.Gremlinq.Providers.Neptune\ExRam.Gremlinq.Providers.Neptune.csproj" />
     <ProjectReference Include="..\..\src\ExRam.Gremlinq.Providers.JanusGraph.AspNet\ExRam.Gremlinq.Providers.JanusGraph.AspNet.csproj" />
     <ProjectReference Include="..\..\src\ExRam.Gremlinq.Providers.JanusGraph\ExRam.Gremlinq.Providers.JanusGraph.csproj" />
-    <ProjectReference Include="..\..\src\ExRam.Gremlinq.Providers.WebSocket\ExRam.Gremlinq.Providers.WebSocket.csproj" />
     <ProjectReference Include="..\ExRam.Gremlinq.Core.Tests\ExRam.Gremlinq.Core.Tests.csproj" />
   </ItemGroup>
